The praxis-card dispatcher moves in beside its leaves
`PraxisCard.tsx` was the second marooned dispatcher: it sat at `components/`
root while the nine skins it dispatches sat in `components/praxisCard/desktop/`.
It is now `components/praxisCard/PraxisCard.tsx`, one level above them --
exactly where `MetataskSeal.tsx` sits relative to `metataskSeal/skins/`.

`ScoreStamp.tsx` is NOT touched. The Kit Map design names it as an offender;
that is backwards. It is already in `praxisCard/scoreStamp/` with its eight
leaves, which SETTLED confirms.

Pure move -- no rendered output changed.
