+++ frontend/src/factions/manifest.ts
@@ -51,7 +51,7 @@ import type { ComponentType } from 'react'
 
 import type { CardProps } from '../components/taskCard/TaskCard'
 import type { SealSkinProps } from '../components/metataskSeal/types'
-import type { ArchetypeProps as PraxisCardProps } from '../components/PraxisCard'
+import type { ArchetypeProps as PraxisCardProps } from '../components/praxisCard/PraxisCard'
 import type { FactionAvatarProps } from '../components/avatar/FactionAvatar'
 import type { SigilVariantProps } from '../components/sigil/FactionSigil'
 import type { CommentComponent } from '../components/comments/shared'
