+++ frontend/src/pages/Home.tsx
@@ -8,7 +8,7 @@ import { useGameConfig } from '../hooks/useGameConfig'
 import { loginWithGoogle, devLogin } from '../api/auth'
 import { computeFactionMultiplier } from '../utils/points'
 import { extractError } from '../utils/errors'
-import PraxisCard from '../components/PraxisCard'
+import PraxisCard from '../components/praxisCard/PraxisCard'
 import TaskCard from '../components/taskCard/TaskCard'
 import ActivityTicker from '../components/home/ActivityTicker'
