+++ frontend/src/pages/characterProfile/archetypes/DefaultProfileBody.tsx
@@ -38,7 +38,7 @@ import { useTranslation } from 'react-i18next'
 import type { BadgeOut } from '../../../api/auth'
 import { badgeArtFor } from '../../../components/badges/badgeArt'
 import CredentialCard from '../../../components/CredentialCard'
-import PraxisCard from '../../../components/PraxisCard'
+import PraxisCard from '../../../components/praxisCard/PraxisCard'
 import TaskCard from '../../../components/taskCard/TaskCard'
 import { useFormFactor } from '../../../hooks/useFormFactor'
 import { factionFill, factionName, isKnownFaction } from '../../../utils/factions'
