+++ frontend/src/pages/characterProfile/archetypes/WowProfileBody.tsx
@@ -46,7 +46,7 @@ import { useState, type ReactNode } from 'react'
 
 import { badgeArtFor } from '../../../components/badges/badgeArt'
 import type { BadgeOut } from '../../../api/auth'
-import PraxisCard from '../../../components/PraxisCard'
+import PraxisCard from '../../../components/praxisCard/PraxisCard'
 import TaskCard from '../../../components/taskCard/TaskCard'
 import {
   WOW_BODY,
