+++ frontend/src/pages/characterProfile/archetypes/profileSkin.tsx
@@ -33,7 +33,7 @@ import { useFormFactor } from '../../../hooks/useFormFactor'
 import type { BadgeOut } from '../../../api/auth'
 import { badgeArtFor } from '../../../components/badges/badgeArt'
 import CredentialCard from '../../../components/CredentialCard'
-import PraxisCard from '../../../components/PraxisCard'
+import PraxisCard from '../../../components/praxisCard/PraxisCard'
 import TaskCard from '../../../components/taskCard/TaskCard'
 import { factionName } from '../../../utils/factions'
 import { mediaUrl } from '../../../utils/media'
