+++ frontend/src/pages/factionDetail/mobileArchetypes/SingularityFactionPage.tsx
@@ -1,7 +1,7 @@
 import { useState, type CSSProperties, type ReactNode } from 'react'
 import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
-import PraxisCard from '../../../components/PraxisCard'
+import PraxisCard from '../../../components/praxisCard/PraxisCard'
 import { factionName, factionDescription } from '../../../utils/factions'
 import { MobileStickyBar } from './shared'
 import type { CharacterOut } from '../../../api/auth'
